[Refactor]: CommunityPostRepository 파일 이동
repository/community -> community/repository 폴더로 이동

Related to: #559
