+++ src/main/java/org/sopt/makers/internal/community/service/CommunityPostServiceUtil.java
@@ -3,7 +3,7 @@ package org.sopt.makers.internal.community.service;
 import org.sopt.makers.internal.community.repository.CommunityPostLikeRepository;
 import org.sopt.makers.internal.domain.community.CommunityPost;
 import org.sopt.makers.internal.exception.NotFoundDBEntityException;
-import org.sopt.makers.internal.repository.community.CommunityPostRepository;
+import org.sopt.makers.internal.community.repository.CommunityPostRepository;
 
 public class CommunityPostServiceUtil {
